Added project-imgs boilerplate
