+++ src/templates/project/project-imgs/index.ts
@@ -1,0 +1,4 @@
+
+import {ProjectImgs} from './project-imgs';
+
+export default ProjectImgs;
