+++ src/templates/project/project-imgs/project-imgs.tsx
@@ -1,0 +1,20 @@
+import React from 'react'
+// import styling libs
+// import local components
+import { Screenshot } from '..'
+
+type Props = {
+    /**
+     * Screenshots of the project (contains mobile img and label)
+     */
+    screenshots: Screenshot[]
+}
+
+/**
+ * Component to render the snapshot / screenshot of the project to be shown..
+ */
+const ProjectImgs: React.FC<Props> = () => {
+    return <></>
+}
+
+export { ProjectImgs }
